Plots locations on a map

The function generates a map plotting all locations, filtered by get.coordinates(). The plot also displays additional information if used by candidates.maps()

simple.map(x, y, cc, color, strings, plot, ratio_string, fq)

Arguments

x

numeric. Latitude of locations filtered by get.coordinates()

y

numeric. Longitude of locations filtered by get.coordinates()

cc

character string. Country code of locations filtered by get.coordinates()

color

character string indicating, which color is assigned to each country code in order.

strings

Character strings in form of regular expression that filter the data frames. The first string is shown on the plot.

plot

logical. If TRUE then the plot will not be printed but saved as .png in the current working directory.

ratio_string

character string. Ratio of occurrences in each polygon returned by top.candidates().

fq

character string. Number of occurrences in the designated polygon and in total.

Value

A plot in the current R session or as .png in the working directory. If directly generated by top(), it displays the first string and the total occurrences. If generated by candidates.maps(), it displays the string, the ratio as a percentage and the number of occurrences in the designated polygon and in total.