Plots locations with specified toponyms on a map

The function calls the simple.map() function to generate a map plotting all locations, filtered by get.coordinates(). The plot also displays additional information if used by candidates.maps(). The data used is downloaded by get.data() and is accessible on the GeoNames download server.

top(
  strings,
  countries = "DE",
  color = rainbow(length(countries)),
  df = TRUE,
  csv = FALSE,
  plot = FALSE,
  ratio_string = "",
  fq = ""
)

Arguments

strings

Character strings in form of regular expression that filter the data frames. The first string is shown on the plot.

countries

Character string with country code abbreviations (check https://www.geonames.org/countries/ for a list of available countries) specifying, the toponyms of which countries are checked.

color

character string indicating, which color is assigned to each country code in order.

df

logical. If TRUE then the filtered data frame will be saved in the global environment.

csv

logical. If TRUE then the filtered data frame will be saved as .csv in the current working directory.

plot

logical. If TRUE then the plot will not be printed but saved as .png in the current working directory.

ratio_string

character string. Ratio of occurrences in each polygon returned by top.candidates(). This should not be specified manually.

fq

character string. Number of occurrences in the designated polygon and in total returned by top.candidates(). This should not be specified manually.

Value

A plot in the current R session or as .png in the working directory. If directly generated by this function top(), it displays the first string and the total occurrences. If generated by candidates.maps(), it displays the string, the ratio as a percentage and the number of occurrences in the designated polygon and in total.

Examples

if (FALSE) {
get.data("DE", save = TRUE)
# saves data for DE in package directory
top("itz$")
## generates and prints a map with all populated places in Germany (default country) ending with "itz" and saves the locations in a data frame in the global environment.


top("^By", "DK", color = "green", df = FALSE, csv = TRUE, plot = TRUE)
## first downloads data for DK and extracts it to the temporary directory
## generates plot with all populated places colored in green in Denmark starting with "By" (case sensitive) and saves it as .png together with the location data as .csv in the working directory.


get.data(c("DE", "PL"), save = TRUE)
# saves for DE and PL in package directory if it's not already there
top(c("itz$", "ice$"), c("DE", "PL"))
# generates plot with all populated places in Germany colored in red and Poland colored in cyan ending with either "itz" or "ice" and saves the locations in a data frame in the global environment.
}